fix: Any(x) where len(x) > 256 && len(x)%32 > 0
The c++ assembly looks correct for Any() which returns a bool by setting
AL, in particular when exiting the non-vectorised loop (.LBB1_9) a final
`testb/setne al` is done.

Unfortunately in the equivalent golang assembly, the full 64 bits of rax
are stored in the return value and as rax is used as a position/counter
in the non-vectorised loop, there may sometimes be non-zero remnants that
get incorporated into the return value. In particular this happens if
this loop is running after the first 256 bools.

The check in the go code specifically tests for not 0, thus returning
the remnants of rax causes incorrect output from the function.

Clearing rax just before setting al seems to fix the issue.